♻️  Refactor graph components in a separate folder
